+++ test/realtime/messages_test.exs
@@ -5,7 +5,7 @@ defmodule Realtime.MessagesTest do
   alias Realtime.Api.Message
   alias Realtime.Database
   alias Realtime.Messages
-  alias Realtime.Repo
+  alias Realtime.Tenants.Repo
 
   setup do
     tenant = Containers.checkout_tenant(run_migrations: true)
